requirements.txt
streamlit
torch
torchvision
pillow
numpy
umap-learn
plotly
tqdm
gdown